follow gh MIT template formatting
The repo is MIT licensed (and will continue to be), but LICENSE was originally created as a local text file with an attribution line to Phil Wang (lucidrains). GitHub does not like this and displays the license as "custom" on GitHub UI.

Re-created using GitHub's license template creator verbatim to ensure things are displayed properly.